+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+*Currently none*
+
+# v1.3.0
 - Fix undefined guild-specific or global nicknames causing blocks to be rejected
     - For `discord.Member` objects, `MemberAdapter` now falls back to `global_name` if
       `nick` is undefined and further to `name` if `global_name` is also undefined